requirements.txt
# Core Framework
Flask==3.0.0
Flask-CORS==4.0.0

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23

# AI/LLM
google-generativeai==0.3.2

# Data Processing
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2

# Environment Variables
python-dotenv==1.0.0

# Utilities
requests==2.31.0
python-dateutil==2.8.2

# Optional for development
werkzeug==3.0.1
